% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMSfileAsMatrix.r
\name{openMSfileAsMatrix}
\alias{openMSfileAsMatrix}
\title{Read raw mass spectrometry data}
\usage{
openMSfileAsMatrix(filename)
}
\arguments{
\item{filename}{Path name of the netCDF, mzData, mzXML or mzML file to read.}
}
\value{
A list includes mz and intensity, "Spec" and information of each epectrum, "DataInfo".
}
\description{
Read raw mass spectrometry data
}
\author{
San-Yuan Wang
}
