% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.mzData.R
\name{write.mzData}
\alias{write.mzData}
\title{Write mass spectrometry data}
\usage{
write.mzData(Spec, DataInfo, filename)
}
\arguments{
\item{Spec}{A dataframe include "mz" and "intensity"}

\item{DataInfo}{A dataframe gererated by openMSfileAsMatrix include "msLevel", "lowMZ", "highMZ", "retentionTime", "peaksCount"}

\item{filename}{Path name of the mzData file to write.}
}
\description{
Write mass spectrometry data
}
\author{
San-Yuan Wang
}
